moved recurring task logic again to fix bugs
